1 fetch (and its properties around components)
