+++ src/components/ProfileRightSide.jsx
@@ -1,12 +1,13 @@
 import React from 'react'
 import {Container} from 'react-bootstrap'
-import Suggestions from './Suggestions'
+//import Suggestions from './Suggestions'
+import SingleSuggestion from './SingleSuggestion'
 
-const ProfileRightSide=()=>{
+const ProfileRightSide=({profilesData})=>{
 
     return(
         <Container className="rounded mt-5" style={{backgroundColor:"yellow",height:500,width:"100%"}}>
-            <Suggestions/>
+            <SingleSuggestion profilesData={profilesData} />
         </Container>
     )
 }
